# lakehoused/lakehoused/api/openapi.yaml
openapi: 3.1.0
info:
  title: Amplifier Daemon API
  version: 1.0.0
  description: |
    REST API for the Amplifier daemon - exposes amplifier CLI functionality as HTTP endpoints.
    
    ## Architecture
    - File-based persistence (JSON)
    - SSE for streaming responses
    - Multi-session concurrency support
    - Stateless session resumption
    
    ## Key Concepts
    - **Session**: A persistent conversation context with Claude
    - **Profile**: Agent configuration with system prompt and tools
    - **Execution**: Single-shot or streaming LLM interaction
    
  contact:
    name: Amplifier Project
    url: https://github.com/microsoft/amplifier
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
...
  /sessions:
    get:
      summary: List sessions
      description: List all sessions with optional filtering
      operationId: listSessions
      tags: [Sessions]
      parameters:
        - name: days_back
          in: query
          description: Number of days to look back
          schema:
            type: integer
            default: 7
            minimum: 1
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionSummary'
              example:
                sessions:
                  - session_id: 123e4567-e89b-12d3-a456-426614174000
                    name: code-review-session
                    created_at: '2025-01-20T10:00:00Z'
                    updated_at: '2025-01-20T10:30:00Z'
                    turns: 5
                    tags: [review, python]

    post:
      summary: Create new session
      description: Create a new conversation session
      operationId: createSession
      tags: [Sessions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            example:
              name: my-session
              profile: code-reviewer
              tags: [review, python]
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /sessions/{session_id}/stream:
    get:
      summary: Stream session events
      description: |
        Server-Sent Events (SSE) endpoint for streaming real-time session updates.
        
        Event types:
        - `message`: Content chunk from LLM
        - `metadata`: Session metadata update
        - `error`: Error occurred
        - `done`: Execution completed
      operationId: streamSession
      tags: [Sessions]
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                message:
                  value: |
                    event: message
                    data: {"content": "Here is my analysis..."}
                    
                    event: metadata
                    data: {"tokens": 150, "duration_ms": 1200}
                    
                    event: done
                    data: {"status": "completed"}
        '404':
          $ref: '#/components/responses/NotFound'

  # Profile Operations
...
components:
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      description: Unique session identifier (UUID)
      schema:
        type: string
        format: uuid
      example: 123e4567-e89b-12d3-a456-426614174000
  schemas:
...
    CreateSessionRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Human-readable session name
        profile:
          type: string
          description: Profile name to use (optional, uses current if not specified)
        tags:
          type: array
          items:
            type: string
          description: Tags for categorization
    SessionSummary:
      type: object
      required: [session_id, name, created_at, updated_at, turns]
      properties:
        session_id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        turns:
          type: integer
          description: Number of conversation turns
        tags:
          type: array
          items:
            type: string
    SessionDetail:
      type: object
      required: [metadata, messages]
      properties:
        metadata:
          $ref: '#/components/schemas/SessionMetadata'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        context:
          type: object
          additionalProperties: true
          description: Additional session context
        config:
          type: object
          additionalProperties: true
          description: Session configuration
    SessionMetadata:
      type: object
      required: [session_id, name, created_at, updated_at, turns]
      properties:
        session_id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        turns:
          type: integer
        total_tokens:
          type: integer
        cost_usd:
          type: number
          format: float
        duration_seconds:
          type: number
          format: float
        tags:
          type: array
          items:
            type: string
    Message:
      type: object
      required: [role, content, timestamp]
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true

    # Execution Models
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              additionalProperties: true
              description: Additional error context
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INVALID_REQUEST
              message: Missing required field 'prompt'
              details:
                field: prompt
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Session not found
              details:
                session_id: 123e4567-e89b-12d3-a456-426614174000
...